+++ docs/source/changelog.rst
@@ -16,10 +16,13 @@ Added
   the new ``client`` keyword argument.  (`#274`_, `@davidt99`_)
 * Support for storing actor exceptions.  (`#156`_)
 * Support for silent :class:`Retries<dramatiq.Retry>`.  (`#295`_)
+* Support for expected exceptions via the ``throws`` actor option.
+  (`#303`_, `@takhs91`_)
 
 .. _#156: https://github.com/Bogdanp/dramatiq/issues/156
 .. _#274: https://github.com/Bogdanp/dramatiq/issues/274
 .. _#295: https://github.com/Bogdanp/dramatiq/issues/295
+.. _#303: https://github.com/Bogdanp/dramatiq/issues/303
 .. _@davidt99: https://github.com/davidt99
 
 Changed